# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reality Compiler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMyAccount
      tags: [me]
      summary: Soft-delete the current user's account
      description: |
        Soft-deletes the user's sessions, archives their marketplace
        listings, anonymises buyer info on past orders, scrubs the
        profile, and revokes active Clerk sessions. A scheduled purge
        job hard-deletes objects and rows after a 30-day grace window.
      responses:
        "200":
          description: Deletion summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/export:
    get:
      operationId: exportMyData
      tags: [me]
      summary: Download a JSON archive of all data tied to the current user
      description: |
        Streams a JSON file containing the user's profile, sessions,
        messages, design outputs, marketplace listings, orders, and
        derived payouts. Returned as an attachment.
      responses:
        "200":
          description: Data archive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDataExport"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/connect-account:
    post:
      operationId: createConnectAccount
      tags: [me]
      summary: Start (or refresh) Stripe Connect Express onboarding
      description: |
        Idempotently creates a Stripe Connect Express account for the
        designer (if one doesn't yet exist) and returns a one-shot
        onboarding URL. The frontend should redirect the browser to
        `onboardingUrl`; Stripe will redirect back to `/payouts` on
        completion.
      responses:
        "200":
          description: Onboarding link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectAccountLink"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Stripe not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/connect-status:
    get:
      operationId: getConnectStatus
      tags: [me]
      summary: Inspect current Stripe Connect account status
      description: |
        Returns the live `charges_enabled` / `payouts_enabled` /
        `details_submitted` flags from Stripe so the UI can prompt the
        designer to finish onboarding when needed.
      responses:
        "200":
          description: Connect status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /me/avatar:
    post:
      operationId: uploadAvatar
      tags: [me]
      summary: Upload a new profile avatar image
      description: |
        Streams an image (PNG, JPEG, or WebP, max 4 MB) directly into
        object storage; the request body is the raw image bytes and the
        `Content-Type` header selects the image type. Returns the updated
        user profile with a hosted `avatarUrl`. The previous avatar
        object is deleted best-effort after a successful upload.
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /me/profile:
    patch:
      operationId: updateMyProfile
      tags: [me]
      summary: Update the authenticated user's public designer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMyProfileInput"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Me:
      type: object
      properties:
        userId:
          type: string
        handle:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
      required:
        - userId
        - handle
    UpdateMyProfileInput:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
          maxLength: 80
        bio:
          type: ["string", "null"]
          maxLength: 500
        avatarUrl:
          description: |
            URL to a hosted image (typically returned by `POST /me/avatar`).
            Set to `null` to clear the avatar.
          type: ["string", "null"]
          maxLength: 1024
    UserDataExport:
      type: object
      additionalProperties: true
      required: [exportedAt, schemaVersion, userId]
      properties:
        exportedAt:
          type: string
          format: date-time
        schemaVersion:
          type: integer
          enum: [1]
        userId:
          type: string
        profile:
          nullable: true
          type: object
          additionalProperties: true
        sessions:
          type: array
          items: { type: object, additionalProperties: true }
        messages:
          type: array
          items: { type: object, additionalProperties: true }
        outputs:
          type: array
          items: { type: object, additionalProperties: true }
        listings:
          type: array
          items: { type: object, additionalProperties: true }
        orders:
          type: array
          items: { type: object, additionalProperties: true }
        payouts:
          type: array
          items: { type: object, additionalProperties: true }
    AccountDeletionSummary:
      type: object
      required: [sessionsDeleted, listingsDeleted, ordersAnonymised, anonId]
      properties:
        sessionsDeleted: { type: integer }
        listingsDeleted: { type: integer }
        ordersAnonymised: { type: integer }
        anonId: { type: string }
...
    ConnectAccountStatus:
      type: string
      enum: [pending, restricted, enabled]
    ConnectAccountLink:
      type: object
      properties:
        accountId:
          type: string
        status:
          $ref: "#/components/schemas/ConnectAccountStatus"
        onboardingUrl:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - accountId
        - status
        - onboardingUrl
        - expiresAt
    ConnectStatus:
      type: object
      properties:
        configured:
          type: boolean
          description: Whether the server has Stripe credentials.
        accountId:
          type: ["string", "null"]
        status:
          oneOf:
            - $ref: "#/components/schemas/ConnectAccountStatus"
            - type: "null"
        chargesEnabled:
          type: boolean
        payoutsEnabled:
          type: boolean
        detailsSubmitted:
          type: boolean
      required:
        - configured
        - accountId
        - status
        - chargesEnabled
        - payoutsEnabled
        - detailsSubmitted
...