# lib/api-zod/src/generated/types/connectAccountLink.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reality Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectAccountStatus } from "./connectAccountStatus";

export interface ConnectAccountLink {
  accountId: string;
  status: ConnectAccountStatus;
  onboardingUrl: string;
  expiresAt: Date;
}