# lib/api-zod/src/generated/types/connectAccountStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reality Compiler API
 * OpenAPI spec version: 0.1.0
 */

export type ConnectAccountStatus =
  (typeof ConnectAccountStatus)[keyof typeof ConnectAccountStatus];

export const ConnectAccountStatus = {
  pending: "pending",
  restricted: "restricted",
  enabled: "enabled",
} as const;